pyproject.toml
[tool.poetry]
name = "wallet"
version = "0.1.0"
authors = ["Kirill Kozhemyakin <kozhemiakink@gmail.com>"]
description = "wallet"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-rest-swagger = "^2.2.0"
mysqlclient = "^2.2.4"
drf-yasg = "^1.21.7"
djangorestframework-jsonapi = "^7.0.2"
django-filter = "^24.2"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.2.2"
mypy = "^1.10.1"
ruff = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
